refactor: updated javascript files to typescript
