+++ apps/web/pages/500.tsx
